Drop support for Minitest
For now, drop support for Minitest to maintain existing behavior and
continue forward momentum.

A follow-up, long-term goal could be to explore how we can improve and
extend Minitest and/or the Rails [test cases][] to better match RSpec.

[test cases]: https://guides.rubyonrails.org/testing.html#a-brief-note-about-test-cases
