+++ README.md
@@ -175,10 +175,6 @@ bin/rails g suspenders:email
 
 ### Testing
 
-#### Minitest
-
-#### RSpec
-
 TODO
 
 ## Contributing
