+++ lib/generators/templates/testing/action_mailer.rb
