+++ lib/generators/templates/testing/chromedriver.rb
