+++ lib/generators/templates/testing/i18n.rb
