+++ lib/generators/templates/testing/shoulda_matchers.rb
